# Speed up PKI installer steps