@frasertweedale 2020-09-30
ACK revoked.  We need to discuss that LDAP profile import code a bit more: https://github.com/freeipa/freeipa/pull/5167#discussion_r497328345