@tiran 2020-09-30
I have changed to PR to handle LDAP cert profile migration differently. It's still not entirely clear if we can drop this migration. It may be required under some circumstances when updating an old cluster to DL1. The migration code now assumes that profiles are migrated when the RA agent cert profile is in LDAP.