@tiran 2020-10-01
I have dropped the commit that removes ``migrate_profiles_to_ldap()``. There is an ongoing discussion how to handle LDAP profile updates and new profiles.